refactor(io): move read and write to io module
