+++ src/index.ts
@@ -1,4 +1,8 @@
-export { Node } from './Node';
-export { Tree } from './Tree';
-export { writeNewick, writeNexus } from './Write';
-export { readNewick } from './Reader';
+export { Node } from './node';
+export { Tree } from './tree';
+export { writeNewick } from './io/writers/newick';
+export { writeNexus } from './io/writers/nexus';
+export { readNewick, readTreesFromNewick } from './io/readers/newick';
+export { readNexus } from './io/readers/nexus';
+export { readPhyloXML } from './io/readers/phyloXML';
+export { readNeXML } from './io/readers/neXML';
